Score, sort and annotate flex-style completions according to match tightness
Up until now, there was no way for completion styles to control
sorting.  This change add such a facility and activated it for the new
"flex" completion style.

The new completion style needs some kind match scoring to be useful,
because "foo" can now match "foobar", "frodo" and "barfromsober".  We
normally want to override the completion table's sorting with the
completion style's, so that "foobar" appears at the top of the
completion list.

* lisp/minibuffer.el (minibuffer-completion-help): Use
completion-style-sort-order and compeltion-style-annotation
properties.
(completion-pcm--hilit-commonality): Propertize completion with
'completion-pcm-commonality-score.
(completion-flx-all-completions): Porpertize completion with
completion-style-sort-order and completion-style-annotation.